add index - partitions, sections, segments
This update should fix a lot of a bugs and make the index visible.

The partitioning feature is still imperfect:
There is currently no blinking cursor to show where you are about to
partition when you click;

The partitioning and sectioning features aren't happy when you try to
partition/section out of order within any one partition or section.

The ability to rename partitions/sections and assign topics exists, but
there is no user interface to do it yet.